backend/llm_validator/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-validator"
version = "1.0.0"
description = "Модуль для проверки юридической легальности изменений в документах"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Legal Tech Team", email = "team@yourise.app"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Legal Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Legal",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "responses>=0.22.0",
]

[project.scripts]
llm-validate = "llm_validator.cli:main"

[project.urls]
Homepage = "https://github.com/yourise/llm-validator"
Repository = "https://github.com/yourise/llm-validator"
Issues = "https://github.com/yourise/llm-validator/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_validator*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
